Move the logic around
