+++ app/src/helpers/getContractInfo.ts
@@ -1,0 +1,76 @@
+import { Contract } from "@prisma/client";
+import { getChainConfig } from ".";
+import getPrisma from "./getPrisma";
+import axios from 'axios';
+
+export default async function getContractInfo(address: string, chain: string): Promise<Contract[] | Error> {
+  const prisma = getPrisma()
+
+  const addressExists: Contract[] = await prisma.contract.findMany({
+    where: {
+      address: address as string,
+      chain: chain,
+    },
+  })
+
+  if (addressExists.length) {
+    return addressExists;
+  }
+
+  const { endpoint, apiKey } = getChainConfig(chain)
+  if (!apiKey) {
+    return new Error('No API key')
+  }
+
+  const { data } = await axios.get(
+    `${endpoint}/api?module=contract&action=getsourcecode&address=${address}&apikey=${apiKey}`
+  )
+  const source = data.result[0].SourceCode
+  const slicedSource = source.substring(1, source.length - 1)
+
+  const contractBase = {
+    address: address,
+    chain: chain,
+    contractName: data.result[0].ContractName,
+    abi: data.result[0].ABI,
+    compilerVersion: data.result[0].CompilerVersion,
+    optimizationUsed: Number(data.result[0].OptimizationUsed),
+    runs: data.result[0].Runs,
+    constructorArguments: data.result[0].ConstructorArguments,
+    evmVersion: data.result[0].EVMVersion,
+    library: data.result[0].Library,
+    licenseType: data.result[0].LicenseType,
+    proxy: data.result[0].Proxy,
+    implementation: data.result[0].Implementation,
+    swarmSource: data.result[0].SwarmSource,
+  }
+  const contracts = []
+
+  try {
+    // When there are more than one contract
+    const parsedSource = JSON.parse(slicedSource)
+    const contractPaths = Object.keys(parsedSource.sources)
+    for (const contractPath of contractPaths) {
+      const eachSource = parsedSource.sources[contractPath].content
+      contracts.push({
+        ...contractBase,
+        contractPath,
+        sourceCode: eachSource,
+      })
+    }
+  } catch (error) {
+    // When there are only one contract
+    contracts.push({
+      ...contractBase,
+      contractPath: '/',
+      sourceCode: data.result[0].SourceCode,
+    })
+  }
+
+  const created: Contract[] = [];
+  for (const contract of contracts) {
+    const newContract: Contract = await prisma.contract.create({ data: contract })
+    created.push(newContract);
+  }
+  return created;
+}
